requirements.txt
selenium
requests
beautifulsoup4
pillow
feedparser
webdriver-manager